Fix/active inactive login logout

+++ backend/src/main/java/com/cinema/backend/repository/UserStatusRepository.java
@@ -4,7 +4,7 @@ import com.cinema.backend.model.UserStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserStatusRepository extends JpaRepository<UserStatus, Short> {
+public interface UserStatusRepository extends JpaRepository<UserStatus, Long> {
 
     // Look up status row by name
     Optional<UserStatus> findByStatusName(String statusName);
